labs/requirements.txt
langchain==0.0.154
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.7.4
streamlit-extras
altair<5
openai
langchain-google-genai==1.0.3
google-generativeai==0.5.2
pytimeparse=1.1.8
